feat: document query factory 구현
