+++ src/types/document.ts
@@ -1,14 +1,3 @@
-export type Document = {
-  documents: DocumentTitle[];
-  totalPageCount: number;
-};
-
-export type DocumentTitle = {
-  id: number;
-  title: string;
-  createdAt: string;
-};
-
 export type NewDocument = {
   title: string;
   fileIds: string[];
@@ -17,20 +6,6 @@ export type NewDocument = {
 
 export type UpdateDocument = NewDocument & {
   documentId: number;
-};
-
-export type DocumentFile = {
-  id: string;
-  name: string;
-  originUrl: string;
-  cdnUrl: string;
-};
-
-export type DocumentDetail = {
-  id: number;
-  title: string;
-  createdAt?: string;
-  files: DocumentFile[];
 };
 
 export type DeleteDocument = {
